# specs/001-medical-agent-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Agent MVP API
  version: 1.0.0
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
  /ingestions/presign:
    post:
      summary: Create pre-signed upload URL
      operationId: createUploadPresign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName, contentType, size]
              properties:
                fileName:
                  type: string
                contentType:
                  type: string
                size:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Presigned URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePresign'

  /ingestions/assets:
    post:
      summary: Confirm uploaded object and register asset
      operationId: completeAssetUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectKey, checksum, diseaseProfileId, reportDate]
              properties:
                objectKey:
                  type: string
                checksum:
                  type: string
                recordId:
                  type: string
                  format: uuid
                diseaseProfileId:
                  type: string
                  format: uuid
                reportDate:
                  type: string
                  format: date
                title:
                  type: string
      responses:
        '200':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAsset'

  /ingestions/parse-jobs:
    post:
      summary: Create async parse job
      operationId: createParseJob
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assetIds, recordId]
              properties:
                assetIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                recordId:
                  type: string
                  format: uuid
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseParseJobCreate'

  /disease-profiles:
    get:
      summary: List disease profiles
      operationId: listDiseaseProfiles
      responses:
        '200':
          description: Disease profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDiseaseProfiles'
    post:
      summary: Create disease profile
      operationId: createDiseaseProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Disease profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDiseaseProfile'

  /report-categories:
    get:
      summary: List report categories
      operationId: listReportCategories
      responses:
        '200':
          description: Report categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReportCategories'
    post:
      summary: Create report category
      operationId: createReportCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Report category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReportCategory'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReportCategory'

  /report-categories/{reportCategoryId}:
    delete:
      summary: Delete report category only when empty
      operationId: deleteReportCategory
      parameters:
        - in: path
          name: reportCategoryId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: onlyIfEmpty
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Report category deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReportCategoryDelete'
        '400':
          description: Invalid category id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReportCategoryDelete'
        '404':
          description: Report category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReportCategoryDelete'
        '409':
          description: Category has associated records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseReportCategoryDelete'

  /disease-profiles/{diseaseProfileId}:
    delete:
      summary: Delete disease profile with all related records and assets
      operationId: deleteDiseaseProfile
      parameters:
        - in: path
          name: diseaseProfileId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: onlyIfEmpty
          required: false
          schema:
            type: boolean
          description: When true, deletion is allowed only if this disease has no associated records.
      responses:
        '200':
          description: Disease profile deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDiseaseProfileDelete'
        '409':
          description: Disease profile has associated records and cannot be deleted in onlyIfEmpty mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDiseaseProfileDelete'
        '404':
          description: Disease profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDiseaseProfileDelete'

  /records/{recordId}:
    get:
      summary: Get record details and latest outputs
      operationId: getRecord
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecord'
    delete:
      summary: Delete one record and related artifacts
      operationId: deleteRecord
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Record deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordDelete'

  /records/{recordId}/source-type:
    patch:
      summary: Update record source type and sync report title
      operationId: updateRecordSourceType
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceType]
              properties:
                sourceType:
                  type: string
      responses:
        '200':
          description: Source type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordSourceTypeUpdate'
        '400':
          description: Invalid record id or source type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordSourceTypeUpdate'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordSourceTypeUpdate'

  /records/{recordId}/analysis:
    get:
      summary: Get cached report analysis or generate and persist one from structured result
      operationId: getRecordAnalysis
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Report analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordAnalysis'
        '400':
          description: Invalid record id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordAnalysis'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordAnalysis'
        '502':
          description: Upstream LLM provider failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordAnalysis'

  /records/{recordId}/trend:
    get:
      summary: Get trend snapshots for current record and previous reports in same disease and category
      operationId: getRecordTrend
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 6
            default: 6
      responses:
        '200':
          description: Trend data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordTrend'
        '400':
          description: Invalid record id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordTrend'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRecordTrend'

  /timeline:
    get:
      summary: Get reverse-chronological timeline by batches
      operationId: getTimeline
      responses:
        '200':
          description: Timeline list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTimeline'

  /timeline/{batchId}:
    get:
      summary: Get batch detail with related materials and outputs
      operationId: getTimelineBatchDetail
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTimelineBatch'